+++ README.md
@@ -10,7 +10,7 @@ ONLY PistacheDev is allowed to modificate the .gitignore file!
 ## Run the application
 1) Make sure that your database is correctly configured and running.
 2) Ensure that your information in the .env file are correct.
-3) Make sure that the config.json file is properly configured.
+3) Make sure that the json/config.json file is properly configured.
 4) Run the application with Nodejs:
 ```sh
 node main
@@ -24,10 +24,10 @@ You just need to run it and wait until the program has finished.
 Refer to the "DATABASE.md" file to set up the database.
 Your database must handle MySQL.
 
-## config.json configuration
-The config.json file contains the official debug configuration.
+## json/config.json
+The json/config.json file contains the official debug configuration.
 For debugging, you can change these settings but you've to revert to the official ones before submitting a pull request.
-If you've edited the config.json, your pull request will be rejected (except if you changed the version setting).
+If you've edited the json/config.json, your pull request will be rejected (except if you changed the version setting).
 
 ## .env configuration
 Create a new file called ".env" in the main folder.
