+++ components/menus/help.js
@@ -1,6 +1,6 @@
 const { EmbedBuilder } = require("discord.js");
 const fs = require("fs");
-const discordPermissions = require("../../permissions.json");
+const discordPermissions = require("../../json/permissions.json");
 
 module.exports =
 {
