+++ components/modals/twitch.js
@@ -20,7 +20,7 @@ module.exports =
                 if (config.length < 1) data = fixMissingConfig(guild);
                 if (newChannel && !guild.channels.cache.get(newChannel)) return interaction.reply(":warning: This channel doesn't exist or the application can't access it!");
 
-                db.query("UPDATE config SET twitch = ? WHERE guild = ?", [!newChannel ? 0 : `${newChannel} 0 0 0`, guild.id], async (err) =>
+                db.query("UPDATE config SET twitch = ? WHERE guild = ?", [!newChannel ? 0 : `${newChannel} 0 0 0 0`, guild.id], async (err) =>
                 {
                     if (err) throw err;
                     let status = ":x: Inactive";
