+++ events/guildCreate.js
@@ -1,5 +1,5 @@
 const { ActionRowBuilder, ButtonBuilder, ButtonStyle, EmbedBuilder } = require("discord.js");
-const config = require("../config.json");
+const config = require("../json/config.json");
 
 module.exports =
 {
