+++ events/guildMemberAdd.js
@@ -1,7 +1,6 @@
-const { EmbedBuilder, AuditLogEvent } = require("discord.js");
-
-// Map for the auto raidmode.
-const newMembers = new Map();
+const { EmbedBuilder } = require("discord.js");
+const { antibots } = require("../guard/antibots");
+const { autoraidmode } = require("../guard/autoraidmode");
 
 module.exports =
 {
@@ -26,35 +25,8 @@ module.exports =
                 if (data[0].memberAdd != 0) await guild.channels.cache.get(data[0].memberAdd).send({ embeds: [embed] });
                 if (data[0].joinRole != 0) member.roles.add(data[0].joinRole);
                 if (data[0].raidmode == 1) member.kick();
-                if (data[0].antibots == 1 && member.user.bot)
-                {
-                    const auditLogs = await guild.fetchAuditLogs({ type: AuditLogEvent.BotAdd, limit: 1 }); // Fetch server logs.
-                    const results = auditLogs.entries;
-                    const log = results.find(entry => entry.targetId == member.id); // Fetch for the latest log with the member ID.
-
-                    if (!log) member.kick(); // For security reasons, kick the bot if no log is available.
-                    if (log && log.executorId != guild.ownerId) member.kick(); // Kick the bot if it wasn't added by the owner.
-                };
-
-                if (data[0].raidmode == 0 && data[0].autoraidmode != 0)
-                {
-                    const [maxMembers, interval] = data[0].autoraidmode.split(" ");
-                    const now = Date.now();
-                    const timestamps = newMembers.get(guild.id) || [];
-                    const filter = timestamps.filter(timestamp => now - timestamp < interval * 1000);
-
-                    // Add the member to the Map.
-                    filter.push(now);
-                    newMembers.set(guild.id, filter);
-
-                    if (filter.length >= maxMembers) // Limit exceeded.
-                    {
-                        db.query("UPDATE config SET raidmode = ? WHERE guild = ?", [1, guild.id], async (err) =>
-                        {
-                            if (err) throw err;
-                        });
-                    };
-                };
+                if (data[0].antibots == 1 && member.user.bot) antibots(member);
+                autoraidmode(data, member);
             });
         }
         catch (err)
