+++ events/messageUpdate.js
@@ -1,6 +1,5 @@
-const { PermissionsBitField, EmbedBuilder, messageLink } = require("discord.js");
-const url = require("../url.json");
-const Perms = PermissionsBitField.Flags;
+const { EmbedBuilder } = require("discord.js");
+const { antilinks } = require("../guard/antilinks");
 
 module.exports =
 {
@@ -10,41 +9,14 @@ module.exports =
         try
         {
             if (!oldMessage.guild) return;
-
-            // Shortcuts.
             const guild = oldMessage.guild;
-            const author = oldMessage.author;
-            const member = oldMessage.member;
-
-            function antiLinks(data)
-            {
-                if (data[0].antilinks == 0) return false;
-                if (data[0].antilinks == 1 && author.bot) return false;
-                if (author.id == client.user.id) return false;
-                if (author.id == guild.ownerId || member.permissions.has(Perms.Administrator)) return false;
-
-                const content = newMessage.content.toLowerCase(); // Convert the text in lower cases.
-                const containLink = url.keywords.some(keyword => content.includes(keyword)); // Search any URL keywords in the text.
-                const isGif = content.includes("tenor.com"); // Check if the sent message 
-
-                if (containLink && !isGif)
-                {
-                    oldMessage.delete().then(() =>
-                    {
-                        oldMessage.channel.send(`:warning: <@${author.id}>, links aren't allowed in this server!`);
-                        return true;
-                    });
-                };
-
-                return false;
-            };
 
             db.query("SELECT * FROM config WHERE guild = ?", [guild.id], async (err, data) =>
             {
                 if (err) throw err;
                 if (data.length < 1) return;
 
-                const containLink = antiLinks(data);
+                const containLink = antilinks(newMessage);
                 if (containLink || data[0].messagesLogs == 0 || oldMessage.author == null || oldMessage.author.bot || oldMessage.content == newMessage.content) return;
 
                 const embed = new EmbedBuilder()
