+++ json/config.json
@@ -1,5 +1,5 @@
 {
-    "version": "3.6.1",
+    "version": "3.7",
     "debug": true,
     "express":
     {
